fix in final solution v1
